requirements.txt
fastapi
uvicorn[standard]
sqlalchemy>=2.0
psycopg2-binary
pydantic>=2
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
pydantic[email]